move match logic into Projections instead of living in Schema, WIP
