+++ processing/src/main/java/org/apache/druid/segment/nested/ScalarNestedCommonFormatColumnSerializer.java
@@ -212,23 +212,29 @@ public abstract class ScalarNestedCommonFormatColumnSerializer<T> extends Nested
     bitmapIndexWriter.setObjectsNotSorted();
     final MutableBitmap[] bitmaps;
     bitmaps = new MutableBitmap[getCardinality()];
-    for (int i = 0; i < bitmaps.length; i++) {
-      bitmaps[i] = indexSpec.getBitmapSerdeFactory().getBitmapFactory().makeEmptyMutableBitmap();
-    }
 
     final IntIterator rows = intermediateValueWriter.getIterator();
     int rowCount = 0;
     while (rows.hasNext()) {
       final int dictId = rows.nextInt();
       encodedValueSerializer.addValue(dictId);
-      bitmaps[dictId].add(rowCount++);
+      MutableBitmap b = bitmaps[dictId];
+      if (b == null) {
+        b = indexSpec.getBitmapSerdeFactory().getBitmapFactory().makeEmptyMutableBitmap();
+        bitmaps[dictId] = b;
+      }
+      b.add(rowCount++);
     }
 
     for (int i = 0; i < bitmaps.length; i++) {
       final MutableBitmap bitmap = bitmaps[i];
-      bitmapIndexWriter.write(
-          indexSpec.getBitmapSerdeFactory().getBitmapFactory().makeImmutableBitmap(bitmap)
-      );
+      if (bitmap == null) {
+        bitmapIndexWriter.write(null);
+      } else {
+        bitmapIndexWriter.write(
+            indexSpec.getBitmapSerdeFactory().getBitmapFactory().makeImmutableBitmap(bitmap)
+        );
+      }
       bitmaps[i] = null; // Reclaim memory
     }
